requirements.txt
customtkinter
opencv-python
numpy
Pillow
PyYAML
requests
tkinterdnd2
google-generativeai
python-dotenv